+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 
 * refactor: replace use of the deprecated ioutil package [#238](https://github.com/Scalingo/go-scalingo/pull/238)
 * build(deps): bump github.com/golang-jwt/jwt/v4 from 4.1.0 to 4.3.0
+* feat(scm-repo-link): Add `SCMRepoLinkList` method [#241](https://github.com/Scalingo/go-scalingo/pull/241)
 
 ## 4.15.1